Example of good scoping
We create a variable dusk inside the logSkyColor() function.
After the if statement, we define a new code block with the {} braces. 
Here we assign a new value to the variable color if the if statement is truthy.

Within the if block, the color variable holds the value 'pink', though outside the if block, 
in the function body, the color variable holds the value 'blue'.

While we use block scope, we still pollute our namespace by reusing the same variable name twice. 
A better practice would be to rename the variable inside the block.