Built-in methods.
.pop() and .push() mutate the array on which they’re called. 
However, there are times that we don’t want to mutate the original array and we can use non-mutating array methods. 
Be sure to check MDN to understand the behavior of the method you are using.

Some arrays methods that are available to JavaScript developers include: 
.join(), .slice(), .splice(), .shift(), .unshift(), and .concat() amongst many others. 
Using these built-in methods make it easier to do some common tasks when working with arrays.