Example on Object Literals
// We use curly braces and an equal sign = {}, to designate an OBJECT LITERAL:
let spaceCat = {}; // empty object
console.log(spaceCat); // Prints {}

/* The object is filled with unordered data organized into key-value pairs. 
A key is like a variable name that points to a location in memory that holds a value. 
The value can be any data type in the language including functions or other objects. 
Key-value pairs inside the object literal are SEPARATED BY COMMAS (,) */