Create 01_Dog_Class.js
You want to create a catalog of all the dogs who belong to a daycare, 
by creating a Dog class that serves as a template for creating new Dog objects. 

An object representing a dog named halley: Dog’s name (a key) is "Halley" (a value) 
and has an age (another key) of 3 (another value). We create the halley object below:

In the Dog class, we use this to set the value of the Dog instance’s name property to the name argument.
Under this.name, we create a property called behavior, which will keep track of the number of times a dog misbehaves. 
The behavior property is always initialized to zero.