remove an elem/whole array containing elem
function justFiltersElem takes a nested array _arr_, and _elem_ as arguments, and returns a new array. 
_elem_ is an element that may or may not be present on one or more of the nested arrays. 
The function uses a for loop to return an array where  _elem_ is removed from all nested arrays.

filteredArray takes a nested array _arr_, and _elem_ as arguments, and returns a new array. 
_elem_ is an element that may or may not be present on one or more of the nested arrays. 
The function uses a for loop to return an array where ANY ARRAY containing _elem_ is removed from _arr_.