myNestedArray has five levels of depth (FCC)
variable _myNestedArray_ is an array.  Modify _myNestedArray_ using any combination of strings, 
numbers, and booleans for data elements so that it has exactly five levels of depth 
(the outer-most array is level 1). Somewhere on the third level, include the string "deep", 
on the fourth level, include the string "deeper", and on the fifth level, include the string "deepest".