Object.keys() to generate an array containing keys
We can generate an array which contains all the keys stored in an object 
using the Object.keys() method and passing in an object as the argument. 
This returns an array with strings representing each property in the object. 
Again, there will be no specific order to the entries in the array.